Updated League to version 2.5
The new version removes several fields from the `LeagueItem` struct
that already appear in `League`. It also has new capabilities including

* support for Teams
* selecting up to 10 Summoners or Teams

With the addition of looking at Leagues by Team, I've added a function
that concatenates a slice of TeamIDs to a comma seperated string. It
works identically to `createSummonerIDString`.

Tests for the new Team interface in `league.go` have been added.