Updated Team to 2.4
The latest version of the Team API allows a client
to request up to 10 summoners at a time. There is also the ability
to search for a team based on their TeamID or up to 10 Teams.

I've also added a function to create comma seperated strings
based on TeamIDs. This is very similar to `createSummonerIDString`
except that it takes a `[]string` as an argument versus a `[]int64`.

A test has been added for `TeamByTeamID`.