Using a new way of creating url parameter strings
After discussing with @TrevorSStone on how best to proceed in #36,
I decided to go with two separate functions for creating parameter
strings. One is focussed on joining a `[]string` and the other
converts a `[]int64` to `[]string` and then joins. This is hidden
behind custom types and a `String()` interface.

I've used this function only in MatchHistory so far but the goal
is to eventually spread it to the other endpoints.